feat(a11y): generalized changes in menu overlay
created FlyoutMenu in order to reuse core functionality in ViewControlsOverlay
