+++ src/components/MenuOverlay/MenuOverlay.scss
@@ -1,87 +1,21 @@
 @import '../../constants/styles';
 @import '../../constants/overlay';
 
-.MenuOverlay {
-  @extend %overlay;
+.MenuItem {
+  @include button-reset;
+  color: var(--text-color);
+  width: 100%;
   display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  
+  flex-direction: row;
+  align-items: center;
+  cursor: pointer;
   border-radius: 4px;
-  background: var(--component-background);
-  box-shadow: 0px 0px 3px 0px var(--document-box-shadow);
-  padding: 6px 0;
 
-  .row {
-    width: 100%;
-    margin: 2px 0;
-    height: 32px;
-    display: flex;
-    align-items: center;
-    padding-left: 12px;
-    padding-right: 18px;
-    cursor: pointer;
-
-    &:hover {
-      background: var(--popup-button-hover);
-    }
-
-    .MenuItem {
-      @include button-reset;
-      color: var(--text-color);
-      width: 100%;
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      cursor: pointer;
-      border-radius: 4px;
-
-      .MenuIcon {
-        margin: 4px;
-      }
-
-      .MenuLabel {
-        margin: 0 4px;
-      }
-    }
+  .MenuIcon {
+    margin: 4px;
   }
 
-  .Button {
-    justify-content: left;
-    width: 100%;
-    margin: 0;
-  }
-
-  &.mobile {
-    position: fixed;
-    width: 100%;
-    left: 0;
-    bottom: 0;
-
-    border-radius: 0px;
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-    border: 0px;
-    box-shadow: 0px 0px 3px 0px var(--document-box-shadow);
-
-    padding-top: 30px;
-    padding-bottom: 30px;
-
-    .swipe-indicator {
-      background: var(--divider);
-      border-radius: 2px;
-      height: 4px;
-      width: 38px;
-      position: absolute;
-      top: 12px;
-      margin-left: auto;
-      margin-right: auto;
-      left: 0;
-      right: 0;
-    }
-
-    .row {
-      padding: 0 24px;
-    }
+  .MenuLabel {
+    margin: 0 4px;
   }
 }
